+++ CHANGELOG.md
@@ -32,12 +32,13 @@ All notable changes to this project will be documented in this file.
 - function `MatDestroy`
 - function `MatPtAP` and `transferMat` for parallel interpolation between non-matching grids, see `examples/hpddm/PtAP-2d-PETSc.edp`
 - preliminary interface to `SVDSolve` from SLEPc to compute singular value decompositions, see `examples/hpddm/mf-2d-SLEPc.edp` or `examples/hpddm/helmholtz-2d-SLEPc-complex.edp`
+- preliminary interface to `NEPSolve` from SLEPc to solve nonlinear eigenvalue problems, see `examples/hpddm/nonlinear-2d-SLEPc-complex.edp`
 - `transpose` parameter when constructing a `Mat` for defining a matrix-free transposed operation
 - interface to `PetscMemoryGetCurrentUsage`
 - add P2b, RT0, RT1 surface FE (P2bS, RT0S, RT1S))
 - add operator interpolate (2d->3d surface)
-- add operator  x = A'*b; where x, b are array and A 2 dim array (full matrix)
--      and genere a error  in case off b'*A  or b'*A expression
+- add operator x = A'\*b; where x, b are array and A 2 dim array (full matrix) and generate an error in case of b'\*A or b'\*A expression
+
 ### Changed
 - moved Htool to its new GitHub location
 - ScaLAPACK and MUMPS are not compiled by PETSc anymore if there is no Fortran compiler
@@ -45,6 +46,7 @@ All notable changes to this project will be documented in this file.
 - PETSc version 3.13.2
 - force `--with-cudac=0` in `make petsc-slepc`, see https://github.com/FreeFem/FreeFem-sources/issues/141
 - change DSL keyword P1dc3dL->P1dcL and P1dc3dS->P1dcS
+
 ### Deprecated
 
 ### Removed